Going more in depth in the shepherds movements, it is actually quite similar to the other entities, the only differance resides in the path construction method which depends on the type of the entity.
Overall, I managed the path interference between multiple shepherds, and highlighting the path in the vue, I did not yet handle the case where two entities meet in opposite directions, in this case, I will recalculate the path.

I've added a class which is called metaData, in the vue, which corresponds to the information that we hold about entities that don't belong to the model such as images, color, ...

For the moment, I've put the status of an entity inside metaata class, I know it is not the best solution, In fact I think we should the status to the model.